fix for Hot reload emits warnings
