+++ .gitignore
@@ -63,3 +63,7 @@ typings/
 
 # next.js build output
 .next
+
+
+# common editor local files 
+.idea/
