+++ package.json
@@ -6,8 +6,8 @@
   "author": "Ivan Aksamentov",
   "license": "MIT",
   "scripts": {
-    "dev": "NODE_ENV=development BABEL_ENV=development node server/index.js",
-    "prod": "NODE_ENV=production BABEL_ENV=production node server/index.js",
+    "dev": "cross-env NODE_ENV=development BABEL_ENV=development node server/index.js",
+    "prod": "cross-env NODE_ENV=production BABEL_ENV=production node server/index.js",
     "clean": "rm -rf .dev/client/** .dev/server/** .prod/client/** .prod/server/**"
   },
   "dependencies": {
@@ -43,13 +43,14 @@
     "babel-eslint": "^8.2.6",
     "babel-loader": "^8.0.0-beta.0",
     "babel-plugin-universal-import": "^3.0.0",
+    "cross-env": "^5.2.0",
     "eslint": "^5.3.0",
     "eslint-loader": "^2.1.0",
     "eslint-plugin-babel": "^5.1.0",
     "eslint-plugin-react": "^7.10.0",
     "react-hot-loader": "^4.3.4",
     "redux-immutable-state-invariant": "^2.1.0",
-    "webpack": "4.16.1",
+    "webpack": "4.16.5",
     "webpack-dev-middleware": "^3.1.3",
     "webpack-hot-middleware": "^2.22.3",
     "webpack-hot-server-middleware": "^0.5.0"
