+++ webpack/server.js
@@ -20,7 +20,8 @@ module.exports = {
   },
 
   module: {
-    rules: [
+      exprContextCritical: false,
+      rules: [
       {
         test: /\.js$/,
         enforce: 'pre',
